requirements.txt
amqp==1.4.9
anyjson==0.3.3
argh==0.26.2
billiard==3.3.0.23
celery==3.1.23
Django==1.9.7
django-celery==3.1.17
django-supervisor==0.3.4
djangorestframework==3.3.3
kombu==3.0.35
Markdown==2.6.6
meld3==1.0.2
pathtools==0.1.2
psycopg2==2.6.1
pytz==2016.4
PyYAML==3.11
requests==2.10.0
supervisor==3.3.0
watchdog==0.8.3
